don't display empty candles calculations in thread visual tweaks

+++ BitcoinCharts/BitcoinChartsWP/Models/Trade.cs
@@ -9,8 +9,8 @@ namespace BitcoinChartsWP.Models
 	public class Trade
 	{
 		public int Id { get; set; }
-		public decimal Amount { get; set; }
-		public decimal Price { get; set; }
+		public double Amount { get; set; }
+		public double Price { get; set; }
 		public DateTimeOffset Timestamp {get;set;}
 	}
 }
